--- gateway/src/ip/Ipv4Or6Address.ts
@@ -1,4 +1,0 @@
-import { Ipv4Address } from './ipv4/Ipv4Address.js';
-import { Ipv6Address } from './ipv6/Ipv6Address.js';
-
-export type Ipv4Or6Address = Ipv4Address | Ipv6Address;
