+++ gateway/src/tests/utils/GatewayClient.ts
@@ -1,8 +1,7 @@
 import { type RawData, WebSocket } from 'ws';
 import { Logger } from 'pino';
 
-import { Ipv4Or6Packet } from '../../ip/Ipv4Or6Packet.js';
-import { initPacket } from '../../ip/packets.js';
+import { initPacket, Ipv4Or6Packet } from '../../ip/ipv4Or6.js';
 
 const DEFAULT_GATEWAY_URL = 'ws://localhost:8080';
 const GATEWAY_URL = process.env.GATEWAY_URL || DEFAULT_GATEWAY_URL;
